+++ src/Lakona.ProjectSystem/Generation/Rendering/Client/UnityClientCodeTemplates.cs
@@ -1109,7 +1109,6 @@ internal static class UnityClientCodeTemplates
 
             private static readonly string[] KnownAnalyzerPackageIds =
             {
-                "Lakona.Rpc.Analyzers",
                 "MemoryPack.Generator",
                 "Microsoft.CodeAnalysis.Common",
                 "Microsoft.CodeAnalysis.CSharp"